Update migration 177 to check for table existence first

+++ migrations/versions/177_add_run_state_eas_folderstatus.py
@@ -15,10 +15,18 @@ import sqlalchemy as sa
 
 
 def upgrade():
+    from inbox.ignition import main_engine
+    engine = main_engine(pool_size=1, max_overflow=0)
+    if not engine.has_table('easfoldersyncstatus'):
+        return
     op.add_column('easfoldersyncstatus', sa.Column('sync_should_run',
                   sa.Boolean(), server_default=sa.sql.expression.true(),
                   nullable=False))
 
 
 def downgrade():
+    from inbox.ignition import main_engine
+    engine = main_engine(pool_size=1, max_overflow=0)
+    if not engine.has_table('easfoldersyncstatus'):
+        return
     op.drop_column('easfoldersyncstatus', 'sync_should_run')
